+++ backends/events/sdl/sdl-events.h
@@ -28,9 +28,6 @@
 
 #include "common/events.h"
 
-// multiplier used to increase resolution for keyboard/joystick mouse
-#define MULTIPLIER 16
-
 // Type names which changed between SDL 1.2 and SDL 2.
 #if !SDL_VERSION_ATLEAST(2, 0, 0)
 typedef SDLKey     SDL_Keycode;
@@ -54,36 +51,17 @@ public:
 	virtual bool pollEvent(Common::Event &event);
 
 	/**
-	 * Resets keyboard emulation after a video screen change
-	 */
-	virtual void resetKeyboardEmulation(int16 x_max, int16 y_max);
-
-	/**
 	 * Emulates a mouse movement that would normally be caused by a mouse warp
 	 * of the system mouse.
 	 */
 	void fakeWarpMouse(const int x, const int y);
 
 protected:
-	/** @name Keyboard mouse emulation
-	 * Disabled by fingolfin 2004-12-18.
-	 * I am keeping the rest of the code in for now, since the joystick
-	 * code (or rather, "hack") uses it, too.
-	 */
-	//@{
-
-	struct KbdMouse {
-		int32 x, y;
-		int16 x_vel, y_vel, x_max, y_max, x_down_count, y_down_count, joy_x, joy_y;
-		uint32 last_time, delay_time, x_down_time, y_down_time;
-		bool modifier;
-	};
-	KbdMouse _km;
-
-	//@}
-
 	/** Scroll lock state - since SDL doesn't track it */
 	bool _scrollLock;
+
+	int _mouseX;
+	int _mouseY;
 
 	/** Joystick */
 	SDL_Joystick *_joystick;
@@ -151,8 +129,6 @@ protected:
 	virtual bool handleJoyButtonUp(SDL_Event &ev, Common::Event &event);
 	virtual bool handleJoyAxisMotion(SDL_Event &ev, Common::Event &event);
 	virtual bool handleJoyHatMotion(SDL_Event &ev, Common::Event &event);
-	virtual void updateKbdMouse();
-	virtual bool handleKbdMouse(Common::Event &event);
 
 #if SDL_VERSION_ATLEAST(2, 0, 0)
 	virtual bool handleJoystickAdded(const SDL_JoyDeviceEvent &event);
@@ -163,17 +139,6 @@ protected:
 #endif
 
 	//@}
-
-	/**
-	 * Update the virtual mouse according to a joystick or game controller axis position change
-	 */
-	virtual bool handleAxisToMouseMotion(int16 xAxis, int16 yAxis);
-
-	/**
-	 * Compute the virtual mouse movement speed factor according to the 'kbdmouse_speed' setting.
-	 * The speed factor is scaled with the display size.
-	 */
-	int16 computeJoystickMouseSpeedFactor() const;
 
 	/**
 	 * Assigns the mouse coords to the mouse event. Furthermore notify the
