+++ backends/events/switchsdl/switchsdl-events.cpp
@@ -97,8 +97,8 @@ void SwitchEventSource::preprocessFingerDown(SDL_Event *event) {
 	// id (for multitouch)
 	SDL_FingerID id = event->tfinger.fingerId;
 
-	int x = _km.x / MULTIPLIER;
-	int y = _km.y / MULTIPLIER;
+	int x = _mouseX;
+	int y = _mouseY;
 
 	if (port == 0 && !ConfMan.getBool("touchpad_mouse_mode")) {
 		convertTouchXYToGameXY(event->tfinger.x, event->tfinger.y, &x, &y);
@@ -141,8 +141,8 @@ void SwitchEventSource::preprocessFingerUp(SDL_Event *event) {
 		}
 	}
 
-	int x = _km.x / MULTIPLIER;
-	int y = _km.y / MULTIPLIER;
+	int x = _mouseX;
+	int y = _mouseY;
 
 	for (int i = 0; i < MAX_NUM_FINGERS; i++) {
 		if (_finger[port][i].id == id) {
@@ -213,8 +213,10 @@ void SwitchEventSource::preprocessFingerMotion(SDL_Event *event) {
 	}
 
 	if (numFingersDown >= 1) {
-		int x = _km.x / MULTIPLIER;
-		int y = _km.y / MULTIPLIER;
+		int x = _mouseX;
+		int y = _mouseY;
+		int xMax = _graphicsManager->getWindowWidth()  - 1;
+		int yMax = _graphicsManager->getWindowHeight() - 1;
 
 		if (port == 0 && !ConfMan.getBool("touchpad_mouse_mode")) {
 			convertTouchXYToGameXY(event->tfinger.x, event->tfinger.y, &x, &y);
@@ -253,14 +255,14 @@ void SwitchEventSource::preprocessFingerMotion(SDL_Event *event) {
 
 			// convert touch events to relative mouse pointer events
 			// Whenever an SDL_event involving the mouse is processed,
-			// _km.x/y are truncated from subpixel precision to regular pixel precision.
-			// Therefore, there's no need here to deal with subpixel precision in _km.x/y.
-			x = (_km.x / MULTIPLIER + (event->tfinger.dx * 1.25 * speedFactor * _km.x_max));
-			y = (_km.y / MULTIPLIER + (event->tfinger.dy * 1.25 * speedFactor * _km.y_max));
+			// _mouseX/Y are truncated from subpixel precision to regular pixel precision.
+			// Therefore, there's no need here to deal with subpixel precision in _mouseX/Y.
+			x = (_mouseX + (event->tfinger.dx * 1.25 * speedFactor * xMax));
+			y = (_mouseY + (event->tfinger.dy * 1.25 * speedFactor * yMax));
 		}
 
-		x = CLIP(x, 0, (int)_km.x_max);
-		y = CLIP(y, 0, (int)_km.y_max);
+		x = CLIP(x, 0, xMax);
+		y = CLIP(y, 0, yMax);
 
 		// update the current finger's coordinates so we can track it later
 		for (int i = 0; i < MAX_NUM_FINGERS; i++) {
@@ -285,8 +287,8 @@ void SwitchEventSource::preprocessFingerMotion(SDL_Event *event) {
 				if (numFingersDownLong >= 2) {
 					// starting drag, so push mouse down at current location (back) 
 					// or location of "oldest" finger (front)
-					int mouseDownX = _km.x / MULTIPLIER;
-					int mouseDownY = _km.y / MULTIPLIER;
+					int mouseDownX = _mouseX;
+					int mouseDownY = _mouseY;
 					if (port == 0 && !ConfMan.getBool("touchpad_mouse_mode")) {
 						for (int i = 0; i < MAX_NUM_FINGERS; i++) {
 							if (_finger[port][i].id == id) {
@@ -346,8 +348,8 @@ void SwitchEventSource::preprocessFingerMotion(SDL_Event *event) {
 }
 
 void SwitchEventSource::convertTouchXYToGameXY(float touchX, float touchY, int *gameX, int *gameY) {
-	int screenH = _km.y_max;
-	int screenW = _km.x_max;
+	int screenH = _graphicsManager->getWindowHeight();
+	int screenW = _graphicsManager->getWindowWidth();
 
 	const int dispW = TOUCHSCREEN_WIDTH;
 	const int dispH = TOUCHSCREEN_HEIGHT;
@@ -369,8 +371,8 @@ void SwitchEventSource::convertTouchXYToGameXY(float touchX, float touchY, int *
 	float dispTouchX = (touchX * (float)dispW);
 	float dispTouchY = (touchY * (float)dispH);
 
-	*gameX = CLIP((int)((dispTouchX - x) / sx), 0, (int)_km.x_max);
-	*gameY = CLIP((int)((dispTouchY - y) / sy), 0, (int)_km.y_max);
+	*gameX = CLIP((int)((dispTouchX - x) / sx), 0, screenW);
+	*gameY = CLIP((int)((dispTouchY - y) / sy), 0, screenH);
 }
 
 void SwitchEventSource::finishSimulatedMouseClicks() {
@@ -388,8 +390,8 @@ void SwitchEventSource::finishSimulatedMouseClicks() {
 					SDL_Event ev;
 					ev.type = SDL_MOUSEBUTTONUP;
 					ev.button.button = simulatedButton;
-					ev.button.x = _km.x / MULTIPLIER;
-					ev.button.y = _km.y / MULTIPLIER;
+					ev.button.x = _mouseX;
+					ev.button.y = _mouseY;
 					SDL_PushEvent(&ev);
 
 					_simulatedClickStartTime[port][i] = 0;
