+++ backends/events/switchsdl/switchsdl-events.h
@@ -45,7 +45,8 @@ private:
 		MAX_TAP_TIME = 250, // taps longer than this will not result in mouse click events
 		MAX_TAP_MOTION_DISTANCE = 10, // max distance finger motion in Vita screen pixels to be considered a tap
 		SIMULATED_CLICK_DURATION = 50, // time in ms how long simulated mouse clicks should be
-	}; // track three fingers per panel
+		MULTIPLIER = 16 // multiplier for sub-pixel resolution
+	};
 
 	typedef struct {
 		int id; // -1: no touch
