+++ backends/events/symbiansdl/symbiansdl-events.h
@@ -23,14 +23,14 @@
 #if !defined(BACKEND_EVENTS_SYMBIAN_SDL_H) && !defined(DISABLE_DEFAULT_EVENTMANAGER)
 #define BACKEND_EVENTS_SYMBIAN_SDL_H
 
-#include "backends/events/sdl/sdl-events.h"
+#include "backends/events/sdl/legacy-sdl-events.h"
 
 #define TOTAL_ZONES 3
 
 /**
  * SDL events manager for Symbian
  */
-class SymbianSdlEventSource : public SdlEventSource {
+class SymbianSdlEventSource : public LegacySdlEventSource {
 public:
 	SymbianSdlEventSource();
